requirements.txt
django==2.1.7
obs-websocket-py==0.5
django-model-utils==4.0.0